fix: 原来 php 给的数据一直都全是字符串

+++ api/upload/page.php
@@ -27,7 +27,7 @@ switch ($_POST['step']) {
 			$ret('你没登录');
 		}
 		$uid = $_SESSION['uid'];
-		$cid = Db::insert([
+		$cid = (int)Db::insert([
 			'uid' => $uid,
 			'fid' => $fid,
 		], 'metas', true);
